move pagemenuedit component to the pagesetedit file to provide more context of usage
