+++ packages/design/src/index.ts
@@ -5,5 +5,4 @@ export { default as Form } from './Form';
 export { defaultPatternComponents } from './Form/components';
 export { default as FormManager, type FormManagerContext } from './FormManager';
 export { defaultPatternEditComponents } from './FormManager/FormEdit/components';
-export { PageMenuEdit } from './FormManager/FormEdit/components/PageMenuEdit';
 export { default as FormRouter } from './FormRouter';
